+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ### 8.10 (2022-10-05)
 - Update dependencies
+- Replace \n new lines with System.lineSeparator()
 
 ### 8.9 (2022-10-04)
 - Update to new major version of json-compare (differences support)